fix: fix NPE in generic HTML

+++ scrapper/src/main/java/backend/academy/scrapper/repository/LinkSignatureRepository.java
@@ -1,12 +1,12 @@
 package backend.academy.scrapper.repository;
 
 import backend.academy.scrapper.model.LinkSignature;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
 public interface LinkSignatureRepository extends JpaRepository<LinkSignature, Long> {
-    Optional<LinkSignature> findByLinkId(Long linkId);
+
+    Optional<LinkSignature> findByLinkId(int linkId);
 }
